Update controller_snapshot_status_check_rate_seconds description to use polling
