+++ packages/forklift-console-plugin/locales/en/plugin__forklift-console-plugin.json
@@ -86,7 +86,7 @@
   "Edit NetworkMap": "Edit NetworkMap",
   "Edit Precopy interval (minutes)": "Edit Precopy interval (minutes)",
   "Edit Provider": "Edit Provider",
-  "Edit Snapshot pooling interval (seconds)": "Edit Snapshot pooling interval (seconds)",
+  "Edit Snapshot polling interval (seconds)": "Edit Snapshot polling interval (seconds)",
   "Edit StorageMap": "Edit StorageMap",
   "Edit URL": "Edit URL",
   "Edit VDDK Init Image": "Edit VDDK Init Image",
@@ -286,7 +286,7 @@
   "Show archived": "Show archived",
   "Show managed": "Show managed",
   "Skip certificate validation": "Skip certificate validation",
-  "Snapshot pooling interval (seconds)": "Snapshot pooling interval (seconds)",
+  "Snapshot polling interval (seconds)": "Snapshot polling interval (seconds)",
   "Something is wrong, the secret was not loaded, please try to reload the page.": "Something is wrong, the secret was not loaded, please try to reload the page.",
   "source": "source",
   "Source Only": "Source Only",
